Make BaseEntityBlock wrenchable, fix datagen IC2 workflow
Datagen with IC2 on the classpath is truly a PITA:
First, IC2 does not work in datagen. It registers its keybinds during mod construction instead of client init, which will fail in datagen because Minecraft options are not initialized. This has been reported by me a couple months ago, however, it remains unfixed to this day.
Furthermore, IC2 is stuck in 2013 and doesn't make use of the Capability system at all, which means you have to implement api interfaces on your classes directly. This can quickly become a huge issue when you don't require IC2, because your mod will crash if the interface class is missing.
So on one hand, I can't use IC2 at runtime in datagen, but on the other hand I need the API classes otherwise my mod will crash. As a workaround, I have implemented a Gradle Artifact Transformer to create a modified IC2 jar with only the api classes locally, to use in datagen.