Recipe system start
This marks the beginning of GTE's reworked recipe management system, which relies on vanilla's existing recipe datapack setup. The idea is to define templates for common recipe types (single-input, multi-output...) to minimize the amount of required code for each recipe type. Combined with our recipe ingredients and outputs, we should be able to cover every possible usage, as well as provide extension methods for addons.